pad_short: totally broken

+++ fillit/fillit.h
@@ -6,14 +6,14 @@
 /*   By: thakala <thakala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/11 16:51:32 by thakala           #+#    #+#             */
-/*   Updated: 2022/01/16 18:02:20 by thakala          ###   ########.fr       */
+/*   Updated: 2022/01/16 18:59:51 by thakala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FILLIT_H
 # define FILLIT_H
 
-unsigned long	pad_short(unsigned short tetrimino, unsigned long index, \
+unsigned long	pad_short(unsigned short tetrimino, unsigned short index, \
 	unsigned char board_width);
 
 #endif
